DataGeneration/requirements.txt
faker>=20.0.0
databricks-sdk>=0.20.0
python-dotenv>=1.0.0

